# An AD compatible besselk implementation for any order or argument (real values)