@heltonmc 2023-05-22
Here is a plot of relative errors for the interesting region...

# master

<img width="603" alt="Screen Shot 2023-05-22 at 2 15 33 PM" src="https://github.com/JuliaMath/Bessels.jl/assets/57471570/16a5aecc-c698-4a41-8312-6efd49dcdbbe">

# this pr

<img width="601" alt="Screen Shot 2023-05-22 at 2 15 50 PM" src="https://github.com/JuliaMath/Bessels.jl/assets/57471570/b4e621ce-1076-49ab-8683-fb31871ea4fa">

The errors along x axis is Arb giving bad results (increasing precision in Arb crashes my computer).

Idk what is going on with the errors in the enzyme stuff will have to track that down next.